// src/app/models/sheet.model.ts
export interface Sheet {
  playername: string;
  username: string;
  ranking: string;
  percentile: string;
  place: string;
  warcount: string;
  nationality: string;
  clanhistory: string;
  cup1on1edition1: string;
  meeting: string;
  cup3on3: string;
  active: boolean;
  ban: boolean;
  lastwar: string;
  fpw: string;
  fpwmax: string;
  fpwmin: string;
  last30days: string;
  last365days: string;
  lastwarpc: string;
  s1wars: string;
  s1fpw: string;
  streak: string;
}